pyproject.toml
[tool.poetry]
name = "syrax"
version = "0.1.0"
description = ""
authors = ["Vinicius Reis <v_reis@outlook.com.br>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.3"
python-decouple = "^3.8"
loguru = "^0.7.2"
witch-doctor = "^1.2.0"
pyfiglet = "^1.0.2"
uvicorn = "^0.32.0"
psycopg = {extras = ["binary"], version = "^3.2.3"}
psycopg-pool = "^3.2.3"


[tool.poetry.group.dev.dependencies]
black = "^24.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
